+++ src/Message/Launch/Validator/Result/LaunchValidationResult.php
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2020 (original work) Open Assessment Technologies SA;
+ * Copyright (c) 2021 (original work) Open Assessment Technologies SA;
  */
 
 declare(strict_types=1);
@@ -27,7 +27,7 @@ use OAT\Library\Lti1p3Core\Message\Payload\MessagePayloadInterface;
 use OAT\Library\Lti1p3Core\Registration\RegistrationInterface;
 use OAT\Library\Lti1p3Core\Util\Result\Result;
 
-class LaunchValidationResult extends Result
+class LaunchValidationResult extends Result implements LaunchValidationResultInterface
 {
     /** @var RegistrationInterface|null */
     private $registration;