+++ src/Security/OAuth2/Validator/RequestAccessTokenValidator.php
@@ -29,12 +29,13 @@ use OAT\Library\Lti1p3Core\Security\Jwt\Parser\ParserInterface;
 use OAT\Library\Lti1p3Core\Security\Jwt\Validator\Validator;
 use OAT\Library\Lti1p3Core\Security\Jwt\Validator\ValidatorInterface;
 use OAT\Library\Lti1p3Core\Security\OAuth2\Validator\Result\RequestAccessTokenValidationResult;
+use OAT\Library\Lti1p3Core\Security\OAuth2\Validator\Result\RequestAccessTokenValidationResultInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
 use Throwable;
 
-class RequestAccessTokenValidator
+class RequestAccessTokenValidator implements RequestAccessTokenValidatorInterface
 {
     /** @var RegistrationRepositoryInterface */
     private $repository;
@@ -63,8 +64,10 @@ class RequestAccessTokenValidator
         $this->parser = $parser ?? new Parser();
     }
 
-    public function validate(ServerRequestInterface $request, array $allowedScopes = []): RequestAccessTokenValidationResult
-    {
+    public function validate(
+        ServerRequestInterface $request,
+        array $allowedScopes = []
+    ): RequestAccessTokenValidationResultInterface {
         $this->reset();
 
         try {